Tokenize input in rep, handle unclosed parens.
In order to handle bad input like "  ,  "  (just a comma) I tokenize the
input in the rep proc and check for an empty token list.

In order to detect unclosed parentheses I added an `eof` marker to the
reader and check for it in the `read_list` proc.  If we run out of tokens
before finding our closing parenthesis then we return a parse error
MalType (which I also just added.)